reverted a bad merge

+++ src/main/java/bfst19/KDTree/BoundingBox.java
@@ -5,29 +5,29 @@ import java.io.Serializable;
 //This class exists because JavaJX BoundingBox is not Serializable
 // and its superclass Bounds only has a 3D constructor
 public class BoundingBox implements Serializable {
-	private float minX;
-	private float minY;
-	private float maxX;
-	private float maxY;
+	private double minX;
+	private double minY;
+	private double maxX;
+	private double maxY;
 
-	public BoundingBox(float minX, float minY, float width, float height) {
+	public BoundingBox(double minX, double minY, double width, double height) {
 		this.minX = minX;
 		this.minY = minY;
 		this.maxX = minX+width;
 		this.maxY = minY+height;
 	}
   
-	public float getMinX() {
+	public double getMinX() {
 		return minX;
 	}
 
-	public float getMinY() {
+	public double getMinY() {
 		return minY;
 	}
 
-	public float getMaxX() { return maxX; }
+	public double getMaxX() { return maxX; }
 
-	public float getMaxY() {
+	public double getMaxY() {
 		return maxY;
 	}
 
@@ -39,8 +39,8 @@ public class BoundingBox implements Serializable {
 				b.getMaxX(), b.getMaxY());
 	}
 
-	public boolean intersects(float minX, float minY,
-							  float maxX, float maxY) {
+	public boolean intersects(double minX, double minY,
+							  double maxX, double maxY) {
 		//if (isEmpty() || w < 0 || h < 0) return false;
 		return (maxX>= getMinX() &&
 				maxY >= getMinY() &&
